+++ CHANGELOG.md
@@ -6,6 +6,8 @@
   [#176](https://github.com/bugsnag/maze-runner/pull/176)
 - Add support for macOS devices using AppiumForMac
   [#177](https://github.com/bugsnag/maze-runner/pull/177)
+- Enforce presence of `Bugsnag-Integrity` header (with `--no-enforce-integrity` option to opt out)
+  [#178](https://github.com/bugsnag/maze-runner/pull/178)
 
 ## Fixes